Create RomanNumeralConverter
This is a simple converter to take roman numerals as an input and convert them to integers. After testing my methodology with simple if-then-else statements and debugging with many print statements, I refined the code to be more memory efficient and reduce runtime.

Note: Does not check if given parameters are valid Roman Numerals. (Perhaps something I will add in the future!)
IDE used: IntelliJ with Java Coretto 16 SDK.